Rename and move Recount command. Imports cs fix
