+++ src/Like/Observers/LikeObserver.php
@@ -11,13 +11,13 @@
 
 namespace Cog\Laravel\Likeable\Like\Observers;
 
+use Cog\Contracts\Likeable\Like\Models\Like as LikeContract;
+use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
 use Cog\Laravel\Likeable\Like\Enums\LikeType;
 use Cog\Laravel\Likeable\Like\Events\ModelWasDisliked;
 use Cog\Laravel\Likeable\Like\Events\ModelWasLiked;
 use Cog\Laravel\Likeable\Like\Events\ModelWasUndisliked;
 use Cog\Laravel\Likeable\Like\Events\ModelWasUnliked;
-use Cog\Contracts\Likeable\Like\Models\Like as LikeContract;
-use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
 
 /**
  * Class LikeObserver.
