+++ src/Providers/LikeableServiceProvider.php
@@ -11,13 +11,13 @@
 
 namespace Cog\Laravel\Likeable\Providers;
 
-use Cog\Laravel\Likeable\Console\LikeableRecountCommand;
 use Cog\Contracts\Likeable\Like\Models\Like as LikeContract;
-use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
 use Cog\Contracts\Likeable\LikeCounter\Models\LikeCounter as LikeCounterContract;
+use Cog\Contracts\Likeable\Services\LikeableService as LikeableServiceContract;
+use Cog\Laravel\Likeable\Console\Commands\Recount;
 use Cog\Laravel\Likeable\Like\Models\Like;
-use Cog\Laravel\Likeable\LikeCounter\Models\LikeCounter;
 use Cog\Laravel\Likeable\Like\Observers\LikeObserver;
+use Cog\Laravel\Likeable\LikeCounter\Models\LikeCounter;
 use Cog\Laravel\Likeable\Services\LikeableService;
 use Illuminate\Support\ServiceProvider;
 
@@ -69,7 +69,7 @@ class LikeableServiceProvider extends ServiceProvider
     {
         if ($this->app->runningInConsole()) {
             $this->commands([
-                LikeableRecountCommand::class,
+                Recount::class,
             ]);
         }
     }
